Add external tag for third-party APIs
This commit adds the pytest.mark.external tag to automatically skip
tests that interact with a third-party API. By default, tests marked
with this tag are skipped. You need to pass the --external flag in order
for them to run. Here's a listing:

    pytest tests/  # will skip everything with @pytest.mark.external
    pytest tests/ --external  # will run everything + @pytest.mark.external
    pytest tests/ -m external  # will ONLY run @pytest.mark.external

In the long-term, the goal is to use an open-source model to check for
these tests and benchmark our stuff.