Initial implementation of textcat task
Not super confident with this implementation as it departed quite a bit
from Prodigy. Instead of getting accept and reject responses, I just ask
the LLM to return POS and NEG.

I also made a few decisions on what the categories would be when using
binary text classification. For now, I'm thinking about using POS and
NEG by default. Not sure if this should be parametrized.