Add llm_task for zero-shot NER (#18)
* Add jinja2 to dependencies

* Add llm_task for zero-shot NER

* Relax dependencies on jinja2 and dotenv

* Make normalizer a subfunction of the registry

I discovered some cases where the choice of the normalizer can affect
an LLM's output. It may be better if this implementation is delegated to
the user via a subfunction in a registry.

* Move component-specific tasks into its own dir

* Add llm_misc registry

I created an extra registry name, misc, to house other things a user
may need. I don't think we need a dedicated normalizer registry, I think
we just need a general-purpose registry that can house easy-to-write
functions similar to spacy.registry.misc

* add first test for NER config

* move loading of envs to backends module

* add ner_predict test method

* allow normalizer to be None (instead of noop)

* fix labels in config

* formatting

* fix labels typing

* fix ambiguity

* Register normalizers to spacy registry

Instead of llm.
The normalizers are generic enough and not really LLM-specific.

* Remove confusion in find_substrings docstring

* Skip the OpenAI test for now to avoid costs

* Add tests to ensure the find_substrings work

* Add tests to check response parser

* Include normalizer in the configuration file

* Remove skip for now to verify if it works non locally

* Add external tag for third-party APIs

This commit adds the pytest.mark.external tag to automatically skip
tests that interact with a third-party API. By default, tests marked
with this tag are skipped. You need to pass the --external flag in order
for them to run. Here's a listing:

    pytest tests/  # will skip everything with @pytest.mark.external
    pytest tests/ --external  # will run everything + @pytest.mark.external
    pytest tests/ -m external  # will ONLY run @pytest.mark.external

In the long-term, the goal is to use an open-source model to check for
these tests and benchmark our stuff.

* Add Github action workflow for third-party tests

This runs our external tests everyday. Similar to spaCy's slow tests

* Make docstrings better

* wording

* extend substrings test for multiple occurrences and check offsets specifically

* fix normalizer bug and add unit tests. remove unnecessary uppercase normalizer.

* formatting

* add assert that normalizer is not None

* add tests for additional & unknown labels

* add multiple-token NER results and unaligned responses

* surface alignment mode to the config and expand unit tests

* raise error when the alignment mode is invalid

* surfice parameters for matching results and add unit tests

---------

Co-authored-by: svlandeg <svlandeg@github.com>