Fix external bug due to saving to disk
The problem is that the query parameter of the backend is initialized
with a function, this causes a config with a function inside, preventing
us to saving properly to disk (invalid JSON/config).

My hacky solution is to make this value pseudo-optional where the
default is None, yet the default value for the query is assigned when
that's the case (query_minichain, query_langchain).