Add explosion-bot workflow
This PR adds the explosion-bot workflow for spacy-llm so that we can run
some CI via an issue comment. One ideal use for this is to have
explosion-bot run a BuildKite pipeline for external tests, something
like:

@explosion-bot please test_external

In order to run external tests (i.e., tests that interact with a
third-party API) in Pull Requests (so that we can double-check if
external tests pass in any machine).

TODOs:
- [ ] Create buildkite pipeline for test_external
- [ ] Add EXPLOSIONBOT_TOKEN and BUILDKITE_SECRET to spacy-llm's Github
      actions