Catch blocking errors from Cohere
Sometimes, Cohere returns a blocking error (e.g., blocked output or
blocked input) because of some violation to their Usage Agreement.  We
want to catch those and give users the option to "ignore" it via the
self._strict argument. The error is often returned as an HTTPError, so
there's no way to catch it from the response.